+++ src/Utility/LinkScanner.php
@@ -23,15 +23,14 @@ use Cake\Event\EventDispatcherTrait;
 use Cake\Event\EventList;
 use Cake\Http\Client;
 use Cake\Http\Client\Response;
-use ErrorException;
 use Exception;
 use Laminas\Diactoros\Stream;
 use LinkScanner\ResultScan;
 use LinkScanner\ScanEntity;
+use LogicException;
 use PHPUnit\Framework\Exception as PHPUnitException;
 use Serializable;
 use Tools\BodyParser;
-use Tools\Exceptionist;
 use Tools\Filesystem;
 
 /**
@@ -162,14 +161,15 @@ class LinkScanner implements Serializable
     /**
      * Internal method to create a lock file
      * @return bool
-     * @throws \ErrorException
+     * @throws \LogicException
      */
     protected function _createLockFile(): bool
     {
-        Exceptionist::isTrue(
-            !$this->getConfig('lockFile') || !file_exists($this->lockFile),
-            __d('link-scanner', 'Lock file `{0}` already exists, maybe a scan is already in progress. If not, remove it manually', $this->lockFile)
-        );
+        if ($this->getConfig('lockFile') && file_exists($this->lockFile)) {
+            throw new LogicException(
+                __d('link-scanner', 'Lock file `{0}` already exists, maybe a scan is already in progress. If not, remove it manually', $this->lockFile)
+            );
+        }
 
         return $this->getConfig('lockFile') && Filesystem::instance()->createFile($this->lockFile);
     }
@@ -236,7 +236,6 @@ class LinkScanner implements Serializable
      * @param string $url Url to scan
      * @param string $referer Referer of this url
      * @return void
-     * @throws \Tools\Exception\KeyNotExistsException
      */
     protected function _recursiveScan(string $url, string $referer = ''): void
     {
@@ -281,7 +280,6 @@ class LinkScanner implements Serializable
      * @param string $url Url to scan
      * @param string $referer Referer of this url
      * @return \Cake\Http\Client\Response|null
-     * @throws \Tools\Exception\KeyNotExistsException
      */
     protected function _singleScan(string $url, string $referer = ''): ?Response
     {
@@ -360,11 +358,13 @@ class LinkScanner implements Serializable
      * @param string|null $filename Filename where to export
      * @return string
      * @see serialize()
-     * @throws \ErrorException
+     * @throws \LogicException
      */
     public function export(?string $filename = null): string
     {
-        Exceptionist::isTrue(!$this->ResultScan->isEmpty(), __d('link-scanner', 'There is no result to export. Perhaps the scan was not performed?'));
+        if ($this->ResultScan->isEmpty()) {
+            throw new LogicException(__d('link-scanner', 'There is no result to export. Perhaps the scan was not performed?'));
+        }
 
         if ($this->getConfig('exportOnlyBadResults')) {
             $this->ResultScan = new ResultScan($this->ResultScan->filter(fn(ScanEntity $item): bool => $item->get('code') >= 400));
@@ -387,7 +387,7 @@ class LinkScanner implements Serializable
      *  - `LinkScanner.resultsImported`: will be triggered when the results have been exported.
      * @param string $filename Filename from which to import
      * @return $this
-     * @throws \ErrorException
+     * @throws \LogicException
      * @noinspection PhpMissingReturnTypeInspection
      */
     public function import(string $filename)
@@ -398,7 +398,7 @@ class LinkScanner implements Serializable
             $instance = unserialize(file_get_contents($filename) ?: '');
         } catch (Exception $e) {
             $message = preg_replace('/file_get_contents\([^)]+\):\s+/', '', $e->getMessage()) ?: '';
-            throw new ErrorException(
+            throw new LogicException(
                 __d('link-scanner', 'Failed to import results from file `{0}` with message `{1}`', $filename, lcfirst($message))
             );
         }
@@ -418,14 +418,16 @@ class LinkScanner implements Serializable
      *
      * Other events will be triggered by `_recursiveScan()` and `_singleScan()` methods.
      * @return $this
-     * @throws \ErrorException
+     * @throws \LogicException
      * @noinspection PhpMissingReturnTypeInspection
      */
     public function scan()
     {
         //Sets the full base url
         $fullBaseUrl = $this->getConfig('fullBaseUrl', Configure::read('App.fullBaseUrl') ?: 'http://localhost');
-        Exceptionist::isUrl($fullBaseUrl, __d('link-scanner', 'Invalid full base url `{0}`', $fullBaseUrl));
+        if (!is_url($fullBaseUrl)) {
+            throw new LogicException(__d('link-scanner', 'Invalid full base url `{0}`', $fullBaseUrl));
+        }
         $this->hostname = get_hostname_from_url($fullBaseUrl);
 
         $this->_createLockFile();